Add null foto

+++ src/models/akunModels.py
@@ -1,6 +1,7 @@
 from email_validator import validate_email, EmailNotValidError
 import phonenumbers
 import bcrypt
+import os
 
 from models.db import mysql
 from models.cdn import bucket
@@ -51,11 +52,12 @@ class Akun:
             self.hashedPassword = bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt())
 
         # UPLOAD FOTO JIKA ADA
-        if foto is not None:
+        foto.seek(0, os.SEEK_END)
+        if foto.tell() != 0:
             blob = bucket.blob(foto.filename)
             blob.upload_from_string(foto.stream.read())
-
             self.gcloudURL = blob.public_url
+
         else:
             self.gcloudURL = None
 
@@ -203,9 +205,15 @@ class Akun:
         return self.gcloudURL
 
     def setFoto(self, foto):
-        # DELETE FOTO
-        blob = bucket.blob(self.gcloudURL.rsplit("/", 1)[-1])
-        blob.delete()
+        # CEK APAKAH FOTO BUKAN FILE KOSONG
+        foto.seek(0, os.SEEK_END)
+        if foto.tell() == 0:
+            raise Exception("File foto tidak ada!")
+            
+        # DELETE FOTO JIKA ADA
+        if self.gcloudURL is not None:
+            blob = bucket.blob(self.gcloudURL.rsplit("/", 1)[-1])
+            blob.delete()
 
         # UPLOAD FOTO
         blob = bucket.blob(foto.filename)
@@ -224,7 +232,10 @@ class Akun:
         cursor.close()
 
     def deleteFoto(self):
-        # DELETE FOTO
+        # CEK APAKAH AKUN MEMILIKI FOTO PROFIL
+        if self.gcloudURL is None:
+            raise Exception(f"Akun {self.username} tidak memiliki foto profil!")
+            
         blob = bucket.blob(self.gcloudURL.rsplit("/", 1)[-1])
         blob.delete()
